requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
pydantic-settings==2.6.1
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==8.3.4
httpx==0.28.0
mysqlclient==2.2.6
python-decouple==3.8
aiofiles==24.1.0
pillow==10.1.0 